% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_biofilm_sequence}
\alias{analyze_biofilm_sequence}
\title{Analyze Biofilm Assembly Sequences}
\usage{
analyze_biofilm_sequence(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\value{
The modified \linkS4class{bsocial} object with assembly path lists in
\code{resultados_analisis$biofilm_gen_paths} and \code{resultados_analisis$biofilm_gr_paths},
and plot functions in \code{graficos}.
}
\description{
Analyze Biofilm Assembly Sequences
}
