% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_growth}
\alias{analyze_growth}
\title{Analyze Growth Metrics}
\usage{
analyze_growth(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\value{
The modified \linkS4class{bsocial} object with a scatter plot stored in
\code{graficos$growth_scatter} and top-10 ranking tables in
\code{resultados_analisis$best_10_ngen} and \code{resultados_analisis$best_10_gr}.
}
\description{
Analyze Growth Metrics
}
