% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_stability}
\alias{analyze_stability}
\title{Analyze Consortium Stability}
\usage{
analyze_stability(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\value{
The modified \linkS4class{bsocial} object with coefficient of variation
data in \code{resultados_analisis$stability_cv_data} and violin plots in
\code{graficos$stability_ngen_plot} and \code{graficos$stability_gr_plot}.
}
\description{
Analyze Consortium Stability
}
