% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-package.R
\docType{package}
\name{bsocialv2-package}
\alias{bsocialv2}
\alias{bsocialv2-package}
\title{bsocialv2: Analysis of Microbial Social Behavior in Bacterial Consortia}
\description{
Provides an S4 class and methods for analyzing microbial social behavior in bacterial consortia. Includes growth parameter extraction, social behavior classification (cooperators/cheaters/neutrals), diversity effect analysis, consortium assembly path finding, and stability analysis via coefficient of variation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Juane99/bsocial}
  \item Report bugs at \url{https://github.com/Juane99/bsocial/issues}
}

}
\author{
\strong{Maintainer}: Juan Emilio Martinez Manjon \email{juanemartinez999@hotmail.com}

Authors:
\itemize{
  \item Maria Garcia Toledo
  \item Jessica Purswani
  \item Rocio Romero Zaliz
}

}
\keyword{internal}
