% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{plot_processed_curves}
\alias{plot_processed_curves}
\title{Plot Preprocessed Growth Curves}
\usage{
plot_processed_curves(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing faceted line plots of
mean growth curves (optical density over time) grouped by experimental condition.
}
\description{
Plot Preprocessed Growth Curves
}
