% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{transform_curated_data}
\alias{transform_curated_data}
\title{Import Pre-processed (Curated) Data}
\usage{
transform_curated_data(.Object, ...)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}

\item{...}{Additional arguments.}
}
\value{
The modified \linkS4class{bsocial} object with \code{datos_procesados} populated
as a data frame containing consortia identifiers, strain presence/absence, and
growth parameters (LogPhase, NGen, GR).
}
\description{
Import Pre-processed (Curated) Data
}
